Fix DbKvs.getRangeOfTimestamps() only returning the first page (#1872)
* Fix DbKvs.getRangeOfTimestamps() only returning the first page

The expectation is to return all results that fall into the range
regardless of the page size. This is how all other KVSs work
at least, and also consistent with getRange().

I'm planning to delete all of this code soon, and just need it to
work in order to execute the intermediate steps of my big sweep
refactor/rewrite.

* Minor edits

* Make checkstyle happy, this time for real